Fix stty size
Docker detects if stdin is a stream for it to activate it's stty settings.
`passthru()` appears to convince docker it is not available.

`proc_open()` seems to let it through fine though, and `stty size` shows
the correct terminal columns/lines. Resizing the terminal window also
updates `stty size` correctly.